core-bot/requirements.txt
# Core-bot dependencies (minimal)
discord.py>=2.3.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
pillow>=10.0.0
